Aula 3 - Autenticacao

+++ gerenciador_tarefas/app/views/tarefas_views.py
@@ -1,14 +1,18 @@
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
-from .forms import TarefaForm
-from .entidades.tarefa import Tarefa
-from .services import tarefa_service
+from ..forms import TarefaForm
+from ..entidades.tarefa import Tarefa
+from ..services import tarefa_service
 
 
+@login_required()
 def listar_tarefas(request):
-    tarefas = tarefa_service.listar_tarefas()
+    tarefas = tarefa_service.listar_tarefas(request.user)
     return render(request, 'tarefas/listar_tarefas.html', {'tarefas': tarefas})
 
 
+@login_required()
 def cadastrar_tarefas(request):
     if request.method == 'POST':
         form_tarefa = TarefaForm(request.POST)
@@ -21,7 +25,8 @@ def cadastrar_tarefas(request):
                 titulo=titulo,
                 descricao=descricao,
                 data_expiracao=data_expiracao,
-                prioridade=prioridade
+                prioridade=prioridade,
+                usuario=request.user,
             )
             tarefa_service.cadastrar_tarefa(tarefa_nova)
             return redirect('listar_tarefas')
@@ -30,8 +35,11 @@ def cadastrar_tarefas(request):
     return render(request, 'tarefas/form_tarefa.html', {'form_tarefa': form_tarefa})
 
 
+@login_required()
 def editar_tarefas(request, id):
     tarefa_bd = tarefa_service.listar_tarefas_id(id)
+    if tarefa_bd.usuario != request.user:
+        return HttpResponse("Não Permitido!")
     form_tarefa = TarefaForm(request.POST or None, instance=tarefa_bd)
     if form_tarefa.is_valid():
         titulo = form_tarefa.cleaned_data['titulo']
@@ -42,15 +50,19 @@ def editar_tarefas(request, id):
             titulo=titulo,
             descricao=descricao,
             data_expiracao=data_expiracao,
-            prioridade=prioridade
+            prioridade=prioridade,
+            usuario=request.user,
         )
         tarefa_service.editar_tarefas(tarefa_bd, tarefa_nova)
         return redirect('listar_tarefas')
     return render(request, 'tarefas/form_tarefa.html', {'form_tarefa': form_tarefa})
 
 
+@login_required()
 def remover_tarefas(request, id):
     tarefa_bd = tarefa_service.listar_tarefas_id(id)
+    if tarefa_bd.usuario != request.user:
+        return HttpResponse("Não Permitido!")
     if request.method == 'POST':
         tarefa_service.remover_tarefas(tarefa_bd)
         return redirect('listar_tarefas')
